+++ tests/test_settings.py
@@ -2088,7 +2088,7 @@ def main() -> None:
     defaults = {}
     setting_ids = set()
 
-    assert data.version == "2.1.2"
+    assert data.version == "2.1.3"
     assert (
         localization["quick_look_card_integration_group"]["en"]
         == "Mod Integration: Quick Look Card"